Made it so a colaborator can have multiple permissions and created logic to create those permissions

+++ src/main/java/com/academy/services/ServiceService.java
@@ -1,16 +1,26 @@
 package com.academy.services;
 
+import com.academy.config.authentication.AuthenticationFacade;
 import com.academy.dtos.service.ServiceMapper;
 import com.academy.dtos.service.ServiceRequestDTO;
 import com.academy.dtos.service.ServiceResponseDTO;
+import com.academy.dtos.service_provider.ServiceProviderRequestDTO;
+import com.academy.dtos.service_provider.ServiceProviderResponseDTO;
+import com.academy.exceptions.AuthenticationException;
+//import com.academy.exceptions.ServiceNotFoundException;
+import com.academy.models.Member;
+import com.academy.models.service.Service;
+import com.academy.models.service.service_provider.ProviderPermissionEnum;
+import com.academy.models.service.service_provider.ServiceProvider;
 import com.academy.exceptions.EntityNotFoundException;
-import com.academy.models.Service;
 import com.academy.models.ServiceType;
 import com.academy.models.Tag;
 import com.academy.repositories.ServiceRepository;
 import com.academy.repositories.ServiceTypeRepository;
+import com.academy.repositories.TagRepository;
 import jakarta.transaction.Transactional;
 
+import java.util.Arrays;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -18,14 +28,28 @@ import java.util.stream.Collectors;
 @org.springframework.stereotype.Service
 public class ServiceService {
 
+    private final ServiceProviderService serviceProviderService;
     private final ServiceRepository serviceRepository;
+    private final TagRepository tagRepository;
     private final ServiceMapper serviceMapper;
+    private final AuthenticationFacade authenticationFacade;
+    private final MemberService memberService;
     private final TagService tagService;
     private final ServiceTypeRepository serviceTypeRepository;
-
-    public ServiceService(ServiceRepository serviceRepository, ServiceMapper serviceMapper, TagService tagService, ServiceTypeRepository serviceTypeRepository) {
+    public ServiceService(ServiceRepository serviceRepository,
+                          TagRepository tagRepository,
+                          ServiceMapper serviceMapper,
+                          ServiceProviderService serviceProviderService,
+                          AuthenticationFacade authenticationFacade,
+                          MemberService memberService,
+                          TagService tagService,
+                          ServiceTypeRepository serviceTypeRepository) {
         this.serviceRepository = serviceRepository;
+        this.tagRepository = tagRepository;
         this.serviceMapper = serviceMapper;
+        this.serviceProviderService = serviceProviderService;
+        this.authenticationFacade = authenticationFacade;
+        this.memberService = memberService;
         this.tagService = tagService;
         this.serviceTypeRepository = serviceTypeRepository;
     }
@@ -33,23 +57,35 @@ public class ServiceService {
     // Create
     @Transactional
     public ServiceResponseDTO create(ServiceRequestDTO dto) {
-        List<Tag> tags = tagService.findOrCreateTagsByNames(dto.tagNames());
-        ServiceType serviceType = serviceTypeRepository.findById(dto.serviceTypeId())
-                .orElseThrow(() -> new EntityNotFoundException(ServiceType.class, dto.serviceTypeId()));
+        Member member = memberService.getMemberByUsername(authenticationFacade.getUsername());
+        Service service = serviceMapper.toEntity(dto, member.getId());
 
-        Service service = serviceMapper.toEntity(dto);
-        service.setServiceType(serviceType);
+        List<Tag> tags = tagService.findOrCreateTagsByNames(dto.tagNames());
         linkServiceToTags(service, tags); // Set up the bidirectional link
 
         Service savedService = serviceRepository.save(service);
-        return serviceMapper.toDto(savedService);
+
+        createOwnerServiceProvider(new ServiceProviderRequestDTO(
+                member.getId(),
+                savedService.getId(),
+                Arrays.asList(ProviderPermissionEnum.values())
+        ));
+        return serviceMapper.toDto(savedService, getPermissionsByProviderUsernameAndServiceId(member.getUsername(), savedService.getId()));
+    }
+    private ServiceProviderResponseDTO createOwnerServiceProvider(ServiceProviderRequestDTO request) {
+        return serviceProviderService.createServiceProvider(request);
     }
 
     // Update
     @Transactional
     public ServiceResponseDTO update(Long id, ServiceRequestDTO dto) {
+        String username = authenticationFacade.getUsername();
         Service existing = serviceRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException(Service.class, id));
+                .orElseThrow(() -> new EntityNotFoundException(Service.class,id));
+
+        List<ProviderPermissionEnum> permissions = getPermissionsByProviderUsernameAndServiceId(username, existing.getId());
+        if(permissions == null || !permissions.contains(ProviderPermissionEnum.UPDATE))
+            throw new AuthenticationException("Member doesn't have permission to update service");
 
         // Remove existing tag associations
         for (Tag tag : new ArrayList<>(existing.getTags())) {
@@ -61,36 +97,42 @@ public class ServiceService {
         List<Tag> newTags = tagService.findOrCreateTagsByNames(dto.tagNames());
         linkServiceToTags(existing, newTags);
 
-        ServiceType serviceType = serviceTypeRepository.findById(dto.serviceTypeId())
-                .orElseThrow(() -> new EntityNotFoundException(ServiceType.class, dto.serviceTypeId()));
-
-        existing.setServiceType(serviceType);
         serviceMapper.updateEntityFromDto(dto, existing);
-
-        return serviceMapper.toDto(serviceRepository.save(existing));
+        serviceRepository.save(existing);
+        return serviceMapper.toDto(existing, permissions);
     }
 
     // Read all
     public List<ServiceResponseDTO> getAll() {
+        String username =  authenticationFacade.getUsername();
         return serviceRepository.findAll()
                 .stream()
-                .map(serviceMapper::toDto)
+                .map(service ->  serviceMapper.toDto(service,
+                        getPermissionsByProviderUsernameAndServiceId(username, service.getId())
+                        ))
                 .collect(Collectors.toList());
     }
 
     // Read one
     public ServiceResponseDTO getById(Long id) {
         Service service = serviceRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException(Service.class, id));
-
-        return serviceMapper.toDto(service);
+                .orElseThrow(() -> new EntityNotFoundException(Service.class,id));
+        String username =  authenticationFacade.getUsername();
+        List<ProviderPermissionEnum> permissions = getPermissionsByProviderUsernameAndServiceId(username, id);
+        if(permissions == null || !permissions.contains(ProviderPermissionEnum.READ))
+            throw new AuthenticationException("Member doesn't have permission to read service");
+        return serviceMapper.toDto(service, getPermissionsByProviderUsernameAndServiceId(username, service.getId()));
     }
 
     // Delete
     @Transactional
     public void delete(Long id) {
+        String username =  authenticationFacade.getUsername();
         Service service = serviceRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException(Service.class, id));
+        List<ProviderPermissionEnum> permissions = getPermissionsByProviderUsernameAndServiceId(username, id);
+        if(permissions == null || !permissions.contains(ProviderPermissionEnum.DELETE))
+            throw new AuthenticationException("Member doesn't have permission to delete service");
 
         service.removeAllTags();
         serviceRepository.delete(service);
@@ -102,4 +144,24 @@ public class ServiceService {
             tag.getServices().add(service);
         }
     }
+
+    public List<ProviderPermissionEnum> getPermissionsByProviderUsernameAndServiceId(String username, Long serviceId){
+        if(!hasServiceProvider(username, serviceId))
+            return null;
+        return serviceProviderService.getPermissionsByProviderUsernameAndServiceId(username, serviceId);
+    }
+    private boolean hasServiceProvider(String username, Long serviceId){
+        return serviceProviderService.existsByServiceIdAndProviderUsername(serviceId, username);
+    }
+    /*
+    @Transactional
+    public ServiceResponseDTO updateMemberPermissions(Long serviceId, Long memberId, List<ProviderPermissionEnum> permissions){
+        ServiceProvider serviceProvider = serviceProviderService.getByServiceIdAndMemberId(serviceId, memberId);
+        serviceProviderService.deleteAllPermissions(serviceProvider);
+        serviceProviderService.addPermissions(serviceProvider, permissions);
+        return getById(serviceId);
+    }
+    */
+
+
 }
